publications_related_diseases_number_and_year
The publications for each related disease. This csv stores the number of publications for each related disease of schizophrenia, as well as the publications per year and the weighted average for the publications of the year. 

For circos plots, you can use each disease as a block. The higher the number of publications, the thicker the block. The higher the weighted average year (publications for each year in average, the brighter the color.  

Calculation example: 1200 publications in 2020, 30 publications in 1980, weighted average = (1200*2020) + (30*1980) / 1230 = 2019 (year).